requirements.txt
pathlib==1.0.1
python-dotenv==0.19.2
requests==2.27.1
python-telegram-bot==13.10
